+++ src/main/kotlin/dev/nybroe/collector/MyBundle.kt
@@ -1,4 +1,4 @@
-package com.github.olivernybroe.collectionsintellij
+package dev.nybroe.collector
 
 import com.intellij.AbstractBundle
 import org.jetbrains.annotations.NonNls