+++ src/main/kotlin/dev/nybroe/collector/services/MyProjectService.kt
@@ -1,0 +1,10 @@
+package dev.nybroe.collector.services
+
+import com.intellij.openapi.project.Project
+import dev.nybroe.collector.MyBundle
+
+class MyProjectService(project: Project) {
+
+    init {
+    }
+}