+++ src/test/resources/inspections/ForeachStatementInspection/foreach-key-value.after.php
@@ -1,0 +1,11 @@
+<?php
+
+$array = [
+    'one' => 'monkey',
+    'two' => 'cow',
+    'tree' => 'snake',
+];
+
+collect($array)->each(function ($animal, $num) {
+    echo "{$num} = {$animal}";
+});