+++ src/test/resources/inspections/ForeachStatementInspection/foreach-key-value.php
@@ -1,0 +1,11 @@
+<?php
+
+$array = [
+    'one' => 'monkey',
+    'two' => 'cow',
+    'tree' => 'snake',
+];
+
+<caret>foreach ($array as $num => $animal) {
+    echo "{$num} = {$animal}";
+}